Rename download to output
